third_party/WebKit/LayoutTests/http/tests/devtools/components/geometry-expected.txt

Tests Geometry utility class


Running: testVectorLength
Testing vector length
Vector length: 7.0711
Vector length: 1.4142
Vector length: 7.0000

Running: testVectorNormalize

Running: testScalarProduct
Scalar Product:18
Scalar Product:-6
Scalar Product:0

Running: testCrossProduct
Cross Product: [-70.0000, 20.0000, 26.0000]
Cross Product: [3.0000, 2.0000, -3.0000]
Cross Product: [0, 0, 0]

Running: testCalculateAngle
Calculate angle: 180.0000
Calculate angle: 141.9840
Calculate angle: 90.0000
Calculate angle: 0

Running: testRadiansToDegrees
deg: 180.0000
deg: 45.0000
deg: 30.0000

Running: testDegreesToRadians
rad: -0.5236
rad: 0
rad: 0.5236
rad: 1.5708
rad: 3.1416

Running: testEulerAnglesFromRotationMatrix
Euler angles: 29.9762 60.0238 45.0200
Euler angles: 60.0238 29.9762 45.0200
Euler angles: 0 19.8684 69.9875

Running: testEulerAnglesToRotate3DString
Transform: rotate3d(0,1,0,0deg) rotate3d(-1,0,0,0deg) rotate3d(0,0,1,0deg)
Transform: rotate3d(0,1,0,1deg) rotate3d(-1,0,0,2deg) rotate3d(0,-0.03489949670250097,0.9993908270190958,3deg)
Transform: rotate3d(0,1,0,-1deg) rotate3d(-1,0,0,-2deg) rotate3d(0,0.03489949670250097,0.9993908270190958,3deg)
Transform: rotate3d(0,1,0,-1deg) rotate3d(-1,0,0,2deg) rotate3d(0,-0.03489949670250097,0.9993908270190958,-3deg)
Transform: rotate3d(0,1,0,0deg) rotate3d(-1,0,0,1deg) rotate3d(0,-0.01745240643728351,0.9998476951563913,2deg)
Transform: rotate3d(0,1,0,1deg) rotate3d(-1,0,0,0deg) rotate3d(0,0,1,2deg)
Transform: rotate3d(0,1,0,1deg) rotate3d(-1,0,0,2deg) rotate3d(0,-0.03489949670250097,0.9993908270190958,0deg)
